new_requirements.txt
absl-py==0.12.0
adal==1.2.6
astroid==2.5.2
astunparse==1.6.3
aztk==0.10.3
azure-batch==5.1.1
azure-common==1.1.27
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-mgmt-batch==5.0.1
azure-mgmt-nspkg==3.0.2
azure-mgmt-storage==3.0.0
azure-nspkg==3.0.2
azure-storage-blob==1.3.1
azure-storage-common==1.4.2
azure-storage-nspkg==3.1.0
bcrypt==3.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
contextlib2==0.6.0.post1
cryptography==3.4.7
csvsort==1.6.1
cycler==0.10.0
dicttoxml==1.7.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
isodate==0.6.0
isort==5.8.0
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
matplotlib
mccabe==0.6.1
msrest==0.6.21
msrestazure==0.6.4
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas
paramiko==2.4.3
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.6.6
PyJWT==2.0.1
pylint==2.7.4
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==3.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy
seaborn==0.11.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-addons
termcolor==1.1.0
tf-slim==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.2
typeguard==2.11.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.4.1
